--- test-results.json
@@ -1,76 +1,0 @@
-{
-  "config": {
-    "configFile": "/Users/dcversus/Documents/GitHub/dcversus.wtf/playwright.config.ts",
-    "rootDir": "/Users/dcversus/Documents/GitHub/dcversus.wtf/tests",
-    "forbidOnly": false,
-    "fullyParallel": false,
-    "globalSetup": null,
-    "globalTeardown": null,
-    "globalTimeout": 0,
-    "grep": {},
-    "grepInvert": null,
-    "maxFailures": 0,
-    "metadata": {},
-    "preserveOutput": "always",
-    "reporter": [
-      [
-        "html",
-        {
-          "outputFolder": "playwright-report",
-          "open": "never"
-        }
-      ],
-      [
-        "json",
-        {
-          "outputFile": "test-results.json"
-        }
-      ]
-    ],
-    "reportSlowTests": {
-      "max": 5,
-      "threshold": 300000
-    },
-    "quiet": false,
-    "projects": [
-      {
-        "outputDir": "/Users/dcversus/Documents/GitHub/dcversus.wtf/test-results",
-        "repeatEach": 1,
-        "retries": 0,
-        "metadata": {},
-        "id": "chromium",
-        "name": "chromium",
-        "testDir": "/Users/dcversus/Documents/GitHub/dcversus.wtf/tests",
-        "testIgnore": [
-          "**/node_modules/**"
-        ],
-        "testMatch": [
-          "**/*.@(spec|test).?(c|m)[jt]s?(x)"
-        ],
-        "timeout": 30000
-      }
-    ],
-    "shard": null,
-    "tags": [],
-    "updateSnapshots": "missing",
-    "updateSourceMethod": "patch",
-    "version": "1.57.0",
-    "workers": 1,
-    "webServer": null
-  },
-  "suites": [],
-  "errors": [
-    {
-      "message": "TypeError: Cannot redefine property: Symbol($$jest-matchers-object)",
-      "stack": "TypeError: Cannot redefine property: Symbol($$jest-matchers-object)\n    at /Users/dcversus/Documents/GitHub/dcversus.wtf/node_modules/@vitest/expect/dist/index.js:680:9"
-    }
-  ],
-  "stats": {
-    "startTime": "2025-12-17T06:38:01.662Z",
-    "duration": 165.90800000000002,
-    "expected": 0,
-    "skipped": 0,
-    "unexpected": 0,
-    "flaky": 0
-  }
-}
